pyproject.toml
[tool.poetry]
name = "bigbroco-server"
version = "0.1.0"
description = ""
authors = ["Jueuunn7 <wndmsrla1234@gmail.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "bigbroco"}]

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.1"
djangorestframework = "^3.15.2"
django-environ = "^0.11.2"
django-storages = "^1.14.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
